Moved blocks generation in docker
